refactor: update import statements to use new scraping_utils module
