+++ tests/document_collector_hub/plugins_test/test_utils.py
@@ -1,7 +1,7 @@
 import unittest
 
 import welearn_datastack.modules.computed_metadata
-import welearn_datastack.utils_.scraping_utils
+import welearn_datastack.modules.scraping_utils
 import welearn_datastack.utils_.text_stat_utils
 
 
@@ -66,6 +66,6 @@ class TestUtils(unittest.TestCase):
     def test_clean_text(self):
         html_text = "<p>Hello, World! Isn't it a beautiful  day?</p>"
         self.assertEqual(
-            welearn_datastack.utils_.scraping_utils.clean_text(html_text),
+            welearn_datastack.modules.scraping_utils.clean_text(html_text),
             "Hello, World! Isn't it a beautiful day?",
         )
