+++ welearn_datastack/plugins/rest_requesters/fao_open_knowledge.py
@@ -24,7 +24,6 @@ from welearn_datastack.exceptions import (
     NoContent,
     NoDescriptionFoundError,
     NotExpectedMoreThanOneItem,
-    PDFFileSizeExceedLimit,
     UnauthorizedLicense,
     UnauthorizedState,
 )
@@ -35,14 +34,14 @@ from welearn_datastack.modules.pdf_extractor import (
     remove_hyphens,
     replace_ligatures,
 )
+from welearn_datastack.modules.scraping_utils import (
+    format_cc_license,
+    remove_extra_whitespace,
+)
 from welearn_datastack.plugins.interface import IPluginRESTCollector
 from welearn_datastack.utils_.http_client_utils import (
     get_http_code_from_exception,
     get_new_https_session,
-)
-from welearn_datastack.utils_.scraping_utils import (
-    format_cc_license,
-    remove_extra_whitespace,
 )
 
 logger = logging.getLogger(__name__)
