+++ welearn_datastack/plugins/rest_requesters/ted.py
@@ -14,9 +14,9 @@ from welearn_datastack.data.source_models.ted import (
 )
 from welearn_datastack.exceptions import NoContent
 from welearn_datastack.modules.computed_metadata import predict_readability
+from welearn_datastack.modules.scraping_utils import clean_return_to_line
 from welearn_datastack.plugins.interface import IPluginRESTCollector
 from welearn_datastack.utils_.http_client_utils import get_new_https_session
-from welearn_datastack.utils_.scraping_utils import clean_return_to_line
 
 logger = logging.getLogger(__name__)
