+++ welearn_datastack/plugins/rest_requesters/uved.py
@@ -26,12 +26,12 @@ from welearn_datastack.exceptions import (
     UnauthorizedState,
 )
 from welearn_datastack.modules.pdf_extractor import get_pdf_content
+from welearn_datastack.modules.scraping_utils import format_cc_license
 from welearn_datastack.plugins.interface import IPluginRESTCollector
 from welearn_datastack.utils_.http_client_utils import (
     get_http_code_from_exception,
     get_new_https_session,
 )
-from welearn_datastack.utils_.scraping_utils import format_cc_license
 
 logger = logging.getLogger(__name__)
