+++ welearn_datastack/plugins/scrapers/conversation.py
@@ -2,21 +2,21 @@ import logging
 import os
 import re
 from datetime import datetime
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Optional
 
 from bs4 import BeautifulSoup, ResultSet  # type: ignore
 from welearn_database.data.models import WeLearnDocument
 
 from welearn_datastack.data.db_wrapper import WrapperRetrieveDocument
+from welearn_datastack.modules.scraping_utils import (
+    clean_return_to_line,
+    extract_property_from_html,
+)
 from welearn_datastack.plugins.interface import IPluginScrapeCollector
 from welearn_datastack.regular_expression import SINGLE_QUOTED_WORD_REGEX
 from welearn_datastack.utils_.http_client_utils import (
     get_http_code_from_exception,
     get_new_https_session,
-)
-from welearn_datastack.utils_.scraping_utils import (
-    clean_return_to_line,
-    extract_property_from_html,
 )
 
 logger = logging.getLogger(__name__)
