+++ welearn_datastack/plugins/scrapers/peerj.py
@@ -1,7 +1,7 @@
 import logging
 import os
 from datetime import datetime, timezone
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Optional
 
 import requests  # type: ignore
 from bs4 import BeautifulSoup, Tag  # type: ignore
@@ -12,14 +12,14 @@ from welearn_database.data.models import WeLearnDocument
 from welearn_datastack.constants import AUTHORIZED_LICENSES
 from welearn_datastack.data.db_wrapper import WrapperRetrieveDocument
 from welearn_datastack.exceptions import NoDOIFoundError, UnauthorizedLicense
+from welearn_datastack.modules.scraping_utils import (
+    clean_return_to_line,
+    extract_property_from_html,
+)
 from welearn_datastack.plugins.interface import IPluginScrapeCollector
 from welearn_datastack.utils_.http_client_utils import (
     get_http_code_from_exception,
     get_new_https_session,
-)
-from welearn_datastack.utils_.scraping_utils import (
-    clean_return_to_line,
-    extract_property_from_html,
 )
 
 logger = logging.getLogger(__name__)
